% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{bcd}
\alias{bcd}
\title{Calculates the box-counting dimension of a polygon}
\usage{
bcd(x, dsn, layer, l = seq(10000, 1e+05, 10000), plot = FALSE, ...)
}
\arguments{
\item{x}{Object of class `sf`}

\item{dsn}{Data source name such as the directory where the file is located}

\item{layer}{Layer name such as the filename of the file to be converted}

\item{l}{Numeric sequence of different cell sizes that will be used to generate the grid. The cell size represents the length of a side of a single square box of the grid}

\item{plot}{Logical; if true, a log-log plot of the number of cells used to cover the polygon vs the cell size will be plotted}

\item{...}{Other arguments passed on to `st_read()` together with `dsn` and `layer`}
}
\value{
A numeric vector of length 1 representing the box-counting dimension of the polygon
}
\description{
Calculates the box-counting dimension of a polygon
}
