% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{count_cells}
\alias{count_cells}
\title{Counts the number of grid cells required to cover the shape polygon}
\usage{
count_cells(grid, f)
}
\arguments{
\item{grid}{The grid as an object of class sp::SpatialPolygons}

\item{f}{The shapefile as an object of class sp::SpatialPolygonsDataFrame}
}
\value{
The number of grid cells of grid required to cover the polygon f
}
\description{
Counts the number of grid cells required to cover the shape polygon
}
