% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_similarity_utilities.R
\name{diff_auc}
\alias{diff_auc}
\title{Calculates step-wise absolute difference between the area under two curves}
\usage{
diff_auc(
  x1_vals,
  x2_vals = x1_vals,
  y1_vals,
  y2_vals,
  unequal = FALSE,
  false_offset = min(c(y1_vals, y2_vals))
)
}
\arguments{
\item{x1_vals}{x-values of the 1st curve}

\item{x2_vals}{x-values of the 2nd curve (equal to `x1_vals` by default)}

\item{y1_vals}{y-values of the 1st curve corresponding to `x1_vals`}

\item{y2_vals}{y-values of the 2nd curve corresponding to `x2_vals`}

\item{unequal}{whether the number of samples in the 1st and 2nd curves are unequal (default is `FALSE`)}

\item{false_offset}{Minimum y-value in `y1_vals` and `y2_vals`}
}
\value{
Numeric value of the absolute difference between the area under the curves
}
\description{
Calculates step-wise absolute difference between the area under two curves
}
