% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_similarity_utilities.R
\name{dist_TA}
\alias{dist_TA}
\title{Calculates distance between two simple polygons or linestrings}
\usage{
dist_TA(p1, p2, p1_theta = 0, type = "polygon")
}
\arguments{
\item{p1}{Coordinates of vertices of the 1st feature in matrix form, such as the one returned by sf::st_coordinates()}

\item{p2}{Coordinates of vertices of the 2nd feature in matrix form, such as the one returned by sf::st_coordinates()}

\item{p1_theta}{Angle of rotation in radians for which the distance is minimum (default is 0)}

\item{type}{The type of feature: 'polygon' for simple polygons (default) or 'linestring' for linestrings}
}
\value{
List containing numeric distance, turning angles with appropriate starting coordinates, harmonised data frame of turning angles with identical x values
}
\description{
Calculates distance between two simple polygons or linestrings
}
