% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_affinity_utilities.R
\name{find_max_extent}
\alias{find_max_extent}
\title{Rotates geometry so that maximum possible extent lies along x-axis}
\usage{
find_max_extent(sf_geom)
}
\arguments{
\item{sf_geom}{sf geometry object whose extent is to be maximized along x-axis}
}
\value{
Rotated sf geometry object such that its maximum possible extent is along x-axis
}
\description{
Rotates geometry so that maximum possible extent lies along x-axis
}
