% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_similarity_utilities.R, R/self_utilities.R
\name{import_sf}
\alias{import_sf}
\title{Wrapper for `sf::st_read()``}
\usage{
import_sf(dsn, layer, ...)

import_sf(dsn, layer, ...)
}
\arguments{
\item{dsn}{Data source name such as the directory where the file is located}

\item{layer}{Layer name such as the filename of the file to be converted}

\item{...}{Other arguments passed on to `st_read()`}
}
\value{
An object of class sf

An object of class sf
}
\description{
Wrapper for `sf::st_read()``

Wrapper for `sf::st_read()``
}
