% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{import_shp}
\alias{import_shp}
\title{Imports an ESRI shapefile containing a single polygon feature}
\usage{
import_shp(directory, shapefile)
}
\arguments{
\item{directory}{The directory where the shapefile is located (without the last trailing slash)}

\item{shapefile}{The filename of the shapefile (without extension) to be converted}
}
\value{
An object of type sp::SpatialPolygonsDataFrame
}
\description{
Imports an ESRI shapefile containing a single polygon feature
}
