% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_utilities.R
\name{import_SVD}
\alias{import_SVD}
\title{Import Spatial Vector Data}
\usage{
import_SVD(x = NULL, dsn, layer, ...)
}
\arguments{
\item{x}{Object of class `sf`}

\item{dsn}{Data source name such as the directory where the file is located}

\item{layer}{Layer name such as the filename of the file to be converted}

\item{...}{Other arguments passed on to `sf::st_read()` together with `dsn` and `layer`}
}
\value{
Object of class `sf`. If CRS of the original data is in Long/Lat, it is transformed to EPSG:3857
}
\description{
Import Spatial Vector Data
}
