% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{linreg.fdim}
\alias{linreg.fdim}
\title{Generates a linear model relating the the number of cells and cell sizes on a logarithmic scale}
\usage{
linreg.fdim(calc_tab)
}
\arguments{
\item{calc_tab}{An object of class fdim}
}
\value{
A linear model of class lm
}
\description{
Generates a linear model relating the the number of cells and cell sizes on a logarithmic scale
}
