% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{mb_dim}
\alias{mb_dim}
\title{Iterates over a loop to generate grids with different cell-sizes, acquire the number of such cells required to cover the polygon and stores them in a matrix}
\usage{
mb_dim(directory = readline(prompt =
  "Please enter the directory without the trailing slash: "),
  shapefile = readline(prompt =
  "Please enter the filename without the extension: "))
}
\arguments{
\item{directory}{The directory name with path where the shapefile is located (without the trailing slash)}

\item{shapefile}{The shapefile in question (without the extension) within the directory}
}
\value{
An object of class fdim which contains a matrix containing cell sizes and corresponding number of those cells required to cover the polygon
}
\description{
Iterates over a loop to generate grids with different cell-sizes, acquire the number of such cells required to cover the polygon and stores them in a matrix
}
\examples{
\dontrun{
calc_tab = mb_dim()
}
\dontshow{
calc_tab = mb_dim(system.file(package = "fdim"), "madhya_pradesh")
}
}
