% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_similarity_utilities.R
\name{merge_x_fill_y}
\alias{merge_x_fill_y}
\title{Harmonise turning angles for two curves with identical x-values}
\usage{
merge_x_fill_y(p1_TA, p2_TA, threshold = 1e-10)
}
\arguments{
\item{p1_TA}{Turning angles of the 1st curve}

\item{p2_TA}{Turning angles of the 2nd curve}

\item{threshold}{Small positive value (default `1e-10`) such that differences lesser than this is ignored}
}
\value{
Data frame with cumulative turning angles for the two curves against identical x-values
}
\description{
Harmonise turning angles for two curves with identical x-values
}
