% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{overlay_grid}
\alias{overlay_grid}
\title{Create grids for polygons supplied as arguments}
\usage{
overlay_grid(shp, cs)
}
\arguments{
\item{shp}{The sp::SpatialPolygonsDataFrame object based on which the grid will be created}

\item{cs}{The cell size of the grid that is to be created. This will be set as the cell width in both x- and y-directions}
}
\value{
An object of class sp::SpatialPolygons which represents the grid as a polygon
}
\description{
Create grids for polygons supplied as arguments
}
