% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{plot.fdim}
\alias{plot.fdim}
\title{Plots a fdim object with the y-axis representing the number of cells required to cover the polygon and the x-axis representing the cell-sizes. It also plots a best-fit line through the points}
\usage{
\method{plot}{fdim}(x, ...)
}
\arguments{
\item{x}{The matrix with the two columns - the cell size and the number of cells of that size required to cover the polygon}

\item{...}{Other arguments}
}
\description{
Plots a fdim object with the y-axis representing the number of cells required to cover the polygon and the x-axis representing the cell-sizes. It also plots a best-fit line through the points
}
\examples{
\dontshow{
calc_tab = mb_dim(system.file(package = "fdim"), "madhya_pradesh")
}
plot(calc_tab)
}
