% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_similarity_utilities.R
\name{plot_TA}
\alias{plot_TA}
\title{Plots turning angles for vertices of a simple polygon or linestring}
\usage{
plot_TA(p, ylim = NULL, col = "black", add = FALSE)
}
\arguments{
\item{p}{List of cumulative normalised distances (`px`) and cumulative turning angles (`py`), as returned by `turning_angle()`}

\item{ylim}{The range of y-values that are to be plotted}

\item{col}{The colour of the turning angles curve that is to be plotted}

\item{add}{Whether to add to the existing plot}
}
\description{
Plots turning angles for vertices of a simple polygon or linestring
}
