% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_similarity_utilities.R
\name{poly_angles}
\alias{poly_angles}
\title{Calculates turning angles for each vertex of a simple polygon or linestring}
\usage{
poly_angles(p_coords, type = "polygon")
}
\arguments{
\item{p_coords}{Coordinates of vertices in matrix form, such as the one returned by `sf::st_coordinates()`}

\item{type}{The type of feature: 'polygon' for simple polygons (default) or 'linestring' for linestrings}
}
\value{
Vector of turning angles in radians
}
\description{
Calculates turning angles for each vertex of a simple polygon or linestring
}
