% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_similarity_utilities.R
\name{poly_distances}
\alias{poly_distances}
\title{Calculates cumulative normalised distances between vertices of a simple polygon or linestring}
\usage{
poly_distances(p_coords, type = "polygon")
}
\arguments{
\item{p_coords}{Coordinates of vertices in matrix form, such as the one returned by sf::st_coordinates()}

\item{type}{The type of feature: 'polygon' for simple polygons (default) or 'linestring' for linestrings}
}
\value{
Vector of cumulative normalised distances in the units of `p_coords`
}
\description{
Calculates cumulative normalised distances between vertices of a simple polygon or linestring
}
