% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_affinity_utilities.R
\name{rotate_about_centroid}
\alias{rotate_about_centroid}
\title{Rotate sf geometry about its centroid}
\usage{
rotate_about_centroid(sf_geom, angle)
}
\arguments{
\item{sf_geom}{sf geometry object}

\item{angle}{Angle (in radians) by which the geometry is to be rotated}
}
\value{
sf geometry object rotated about its centroid by the desired angle
}
\description{
Rotate sf geometry about its centroid
}
