% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_similarity_utilities.R
\name{shifter}
\alias{shifter}
\title{Circular shifting of a data frame by rows with identical first and last rows}
\usage{
shifter(x, n = 1)
}
\arguments{
\item{x}{Data frame with identical first and last rows}

\item{n}{Step-size by which the shift is to be made; n = 1 by default}
}
\value{
Data frame `x` with `n` rows shifted circularly and identical first and last rows
}
\description{
Circular shifting of a data frame by rows with identical first and last rows
}
