% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{similarity_linestring}
\alias{similarity_linestring}
\title{Calculates similarity as a metric between two linestrings}
\usage{
similarity_linestring(l1, l2, plot = TRUE)
}
\arguments{
\item{l1}{1st Linestring as a simple feature, as an `sf` object}

\item{l2}{2nd Linestring as a simple feature, as an `sf` object}

\item{plot}{Whether to plot the turning angles}
}
\value{
Similarity value along with the configuration for which this was obtained
}
\description{
Calculates similarity as a metric between two linestrings
}
\examples{
l1 = sf::st_sf(sf::st_sfc(sf::st_linestring(cbind(1:100, sin(sort(runif(100)*10))))))
l2 = sf::st_sf(sf::st_sfc(sf::st_linestring(cbind(1:100, cos(sort(runif(100)*10))))))
similarity_linestring(l1, l2)$dist

}
