% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{similarity_polygon}
\alias{similarity_polygon}
\title{Calculates similarity as a metric between two simple polygons}
\usage{
similarity_polygon(p1, p2, plot = TRUE)
}
\arguments{
\item{p1}{1st Polygon as a simple feature, as an `sf` object}

\item{p2}{2nd Polygon as a simple feature, as an `sf` object}

\item{plot}{Whether to plot the turning angles}
}
\value{
Similarity value along with the configuration for which this was obtained
}
\description{
Calculates similarity as a metric between two simple polygons
}
\examples{
bhutan = rnaturalearth::ne_countries(country = "bhutan", scale = "medium")
switzerland = rnaturalearth::ne_countries(country = "switzerland", scale = "medium")
similarity_polygon(bhutan, switzerland)$dist

}
