% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{summary.fdim}
\alias{summary.fdim}
\title{Prints the fractal dimension of the polygon}
\usage{
\method{summary}{fdim}(object, ...)
}
\arguments{
\item{object}{An object of class fdim}

\item{...}{Other arguments}
}
\description{
Prints the fractal dimension of the polygon
}
\examples{
\dontshow{
calc_tab = mb_dim(system.file(package = "fdim"), "madhya_pradesh")
}
summary(calc_tab)
}
