% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_similarity_utilities.R
\name{turning_angle}
\alias{turning_angle}
\title{Cumulative turning angles vs. cumulative normalised length of a simple polygon or linestring}
\usage{
turning_angle(p_coords, type = "polygon")
}
\arguments{
\item{p_coords}{Coordinates of vertices in matrix form, such as the one returned by sf::st_coordinates()}

\item{type}{The type of feature: 'polygon' for simple polygons (default) or 'linestring' for linestrings}
}
\value{
List of vectors `px` (cumulative normalised distances) and `py` (corresponding cumulative turning angles)
}
\description{
Cumulative turning angles vs. cumulative normalised length of a simple polygon or linestring
}
