% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_affinity_utilities.R
\name{x_extent}
\alias{x_extent}
\title{Calculate extent along x-axis}
\usage{
x_extent(sf_geom)
}
\arguments{
\item{sf_geom}{sf geometry object whose extent along x-axis is to be calculated}
}
\value{
Length of the extent along x-axis
}
\description{
Calculate extent along x-axis
}
