% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_affinity_utilities.R
\name{y_extent}
\alias{y_extent}
\title{Calculate extent along y-axis}
\usage{
y_extent(sf_geom)
}
\arguments{
\item{sf_geom}{sf geometry object whose extent along y-axis is to be calculated}
}
\value{
Length of the extent along y-axis
}
\description{
Calculate extent along y-axis
}
