% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self.R
\name{bcd}
\alias{bcd}
\alias{bcd.self_similarity}
\alias{bcd.self_affinity}
\title{Calculates the box-counting dimension of a polygon}
\usage{
bcd(k, type, l = seq(10000, 1e+05, 10000), plot = FALSE)

\method{bcd}{self_similarity}(k, type = "s", l = seq(10000, 1e+05, 10000), plot = FALSE)

\method{bcd}{self_affinity}(k, type = "a", l = seq(10000, 1e+05, 10000), plot = FALSE)
}
\arguments{
\item{k}{Spatial Vector data as an `sf` object}

\item{type}{One of {'s', 'a'} (representing self-similarity and self-affinity respectively) that is to be calculated}

\item{l}{Numeric; represents the length of a side of a single square box of the grid to calculate box-counting dimension}

\item{plot}{Logical; if true, a log-log plot of the number of cells used to cover the polygon vs the cell size will be plotted}
}
\value{
A numeric representing the box-counting dimension of the spatial feature, `k`.
}
\description{
Calculates the box-counting dimension of a polygon

Calculates self-similar box-counting dimensions

Calculates self-affine box-counting dimensions
}
