% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_similarity_utilities.R
\name{count_cells}
\alias{count_cells}
\title{Counts the number of grid cells required to cover the shape polygon}
\usage{
count_cells(grid, f)
}
\arguments{
\item{grid}{The grid as an object of class sfc}

\item{f}{sf object that the grid is based on}
}
\value{
The number of grid cells of grid required to cover the polygon f
}
\description{
Counts the number of grid cells required to cover the shape polygon
}
